+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+ - New plugin Cart
+
+### Removed
+ - Plugin Basket
+ 
 ## [2.8.1] - 2017-12-19
 ### Fixed
  - Form with paragraph throws RNG error
@@ -292,6 +299,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Plugin LinkList creates and displays local variable ``linklist``.
 - System messages from previous request are now marked via class.
 
+[Unreleased]: https://bitbucket.org/igwr/cms/compare/dev..master
 [2.8.1]: https://bitbucket.org/igwr/cms/compare/v2.8.1..v2.8.0
 [2.8.0]: https://bitbucket.org/igwr/cms/compare/v2.8.0..v2.7.2
 [2.7.2]: https://bitbucket.org/igwr/cms/compare/v2.7.2..v2.7.1